refactor did you know section
